sw_services: xilloader: Fixes for MISRA C violation Rule 10.3
MISRA C Rule 10.3 States: The value of an expression shall not be
assigned to an object with a narrower essential type or of a different
essential type category

This patch fixes MISRA C Rule 10.3 violation by using the proper type
variables where required or typecasting if necessary


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilloader/src/xloader.c
@@ -1009,12 +1009,12 @@ static int XLoader_GetChildRelation(u32 ChildImgID, u32 ParentImgID, u32 *IsChil
 		}
 
 		if (TempParentImgID == 0U) {
-			*IsChild = FALSE;
+			*IsChild = (u32)FALSE;
 			break;
 		}
 
 		if (TempParentImgID == ParentImgID) {
-			*IsChild = TRUE;
+			*IsChild = (u32)TRUE;
 			break;
 		}
 		TempImgID = TempParentImgID;
@@ -1065,7 +1065,7 @@ static int XLoader_InvalidateChildImgInfo(u32 ParentImgID, u32 *ChangeCount)
 			ImageInfoTbl.Count--;
 			(*ChangeCount)++;
 			if (ImageInfoTbl.IsBufferFull == TRUE) {
-				ImageInfoTbl.IsBufferFull = FALSE;
+				ImageInfoTbl.IsBufferFull = (u8)FALSE;
 			}
 		}
 		else if (ImageInfoTbl.TblPtr[Index].ImgID != XLOADER_INVALID_IMG_ID) {
@@ -1165,7 +1165,7 @@ int XLoader_StoreImageInfo(XLoader_ImageInfo *ImageInfo)
 
 	ImageEntry = XLoader_GetImageInfoEntry(ImageInfo->ImgID);
 	if (ImageEntry == NULL) {
-		ImageInfoTbl.IsBufferFull = TRUE;
+		ImageInfoTbl.IsBufferFull = (u8)TRUE;
 		Status = XPlmi_UpdateStatus(XLOADER_ERR_IMAGE_INFO_TBL_OVERFLOW, 0);
 		goto END;
 	}
